TwoSum

+++ src/BasicJava/Basic/CelsiusToFahreheit.java
@@ -8,7 +8,7 @@ public class CelsiusToFahreheit {
         System.out.println("Enter the temperature !");
         double temperature = scanner.nextDouble();
 
-        double celsius = (9/5) * temperature +32;
+        double celsius = ((double) 9 /5) * temperature +32;
         System.out.println("Convert temperature celsius to fahrenheit " + celsius);
     }
 }
